+++ tenant/angular4/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "vsicons.presets.angular": true
+}
