--- tenant/angular4/src/app/repo/repo.component.html
@@ -1,118 +1,0 @@
-<div class="d3-design" #design></div>
-<md-card>
-  <md-card-title>Create Repository</md-card-title>
-  <!--<md-card-subtitle>md-card is a content container component that conforms to the spec of a Material Design card.</md-card-subtitle>-->
-  <md-card-content>
-    <md-card>
-      <md-card-content>
-
-
-                       <md-radio-group name="repotype" [(ngModel)]="repo.islimit"  fxLayout="column">
-                            <div fxLayout="column">
-                                <md-radio-button name="privateRbx1" [value]=true  ngDefaultControl>
-                                    <label for="privateRbx1" class="">
-                                    <span >Private</span>
-                                    <span >- only authorized users can download your files.</span>
-                                    </label></md-radio-button>
-                                <md-radio-button name="publicRbx2"  [value]=false ngDefaultControl>
-                                    <label for="publicRbx2">
-                                    <span >Public</span>
-                                    <span>- anyone can download your files.</span>
-                                    </label>
-                                </md-radio-button>
-                            </div>
-                       </md-radio-group>
-
-    </md-card-content>
-    <md-card>
-    </md-card>
-  <!--<form [formGroup]="form">-->
-
-      <md-card-content>
-          
-                 <div fxLayout="column" class="mb-1" fxLayoutWrap="wrap">
-                    <md-input-container  style="width: 50%;" [(ngModel)]="repo.name" ngDefaultControl>
-                    <input mdInput placeholder="Name"   [formControl]="form.controls['fname']"   type="Cannot be changed once the repository is created." >
-                    </md-input-container>
-                  
-                    <small style ="margin-left:5px" *ngIf="form.controls['fname'].hasError('required') && form.controls['fname'].touched" class="mat-text-warn">You must include a name.</small>
-                    <small *ngIf="form.controls['fname'].hasError('minlength') && form.controls['fname'].touched" class="mat-text-warn">Your  name must be at least 5 characters long.</small>
-                    <small *ngIf="form.controls['fname'].hasError('maxlength') && form.controls['fname'].touched" class="mat-text-warn">Your  name cannot exceed 10 characters.</small>
-                   
-                </div>
-
-                <div fxLayout="column"> Type
-
-
-                    <md-radio-group name="repotype" [(ngModel)]="repo.repo_type"  fxLayout="column">
-                        <md-radio-button  name="repotype" [value]="0">
-                                        <label for="typeRbx" class="" >
-                                        <span >Docker Images</span>
-                                        <span >- An Docker Image is a read-only template with instructions for creating a Docker container.</span>
-                                        </label>
-                        
-                            </md-radio-button>
-                            <div *ngIf="!repo.repo_type" class="pr-1 hide-on-open" class="repo-input-button" style="margin: 10px">
-                                            <!--<label for="typeRbx" class="" >
-                                            </label>-->
-                                                  <button md-raised-button >                                                
-                                                      <span >docker login -u {{username}} -p &lt;API_KEY	&gt; -e &lt;EMAIL	&gt; {{webside_url}}</span>
-                                                </button>
-
-                            </div>
-                        <md-radio-button  name="repotype" [value]="1">
-                        <label for="typeBFRbx">
-                                        <span >Binary Files</span>
-                                        <span>- any kinds of file types.</span>
-                            </label>
-                        </md-radio-button>
-                    </md-radio-group>
-                    <!--<p>Your select is: {{repo.repo_type}}</p>-->
-
-                 </div>
-                <br>
-                <hr>
-                <br>
-                <div fxLayout="column"  class="mb-2" >
-                    <md-input-container  style="width: 100%;" [(ngModel)]="repo.Description" ngDefaultControl>
-                        <input mdInput placeholder="Description (Optional)"  [formControl]="form.controls['desc']" type="Description (Optional)">
-                    </md-input-container>
-                    <small *ngIf="form.controls['desc'].hasError('minlength') && form.controls['desc'].touched" class="mat-text-warn">Your description must be at least 5 characters long.</small>
-                    <small *ngIf="form.controls['desc'].hasError('maxlength') && form.controls['desc'].touched" class="mat-text-warn">Your description cannot exceed 100 characters.</small>
-
-                </div>
-       
-         
-        
-        </md-card-content>
-
-        <md-card-content>
-            <br>
-            <md-checkbox [(ngModel)]="repo.ishavekey" >Repository files using Binary's public/private key pair.</md-checkbox>
-        </md-card-content>
-
-       <md-card-actions>
-            <br>
-             <div *ngIf="!repo.name||form.controls['fname'].hasError('minlength')||form.controls['fname'].hasError('maxlength')||repo.repo_type===-1" class="pr-1 hide-on-open">
-            <button md-raised-button disabled class="mr-1" value="Create" >Create</button>
-            <button md-raised-button class="mr-1">Cancel</button>
-
-             </div>
-             <div *ngIf="repo.name&&!form.controls['fname'].hasError('minlength')&&!form.controls['fname'].hasError('maxlength')&&repo.repo_type!=-1" class="pr-1 hide-on-open">
-            <button md-raised-button color="primary" class="mr-1" value="Create" (click)="onCreateButtonClick(repo)" >Create</button>
-            <button md-raised-button class="mr-1">Cancel</button>
-             </div>
-
-            <div *ngIf="isCreate" class="pr-1 hide-on-open" class="repo-input-button" style="margin: 10px">
-                  <md-card-footer>
-                    <md-progress-bar mode="indeterminate"></md-progress-bar>
-                  </md-card-footer>
-             </div>
-               
-      </md-card-actions>
-    </md-card>
-  
-  </md-card-content>
-    <!--</form>-->
-
-</md-card>
