--- tenant/angular4/src/app/repo/repo.component.scss
@@ -1,62 +1,0 @@
-/*
-Copyright 2016 - 2017 Huawei Technologies Co., Ltd. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-@import "../../assets/styles/scss/components/chat";
-
-:host {
-  margin-left: -5px;
-  margin-right: -5px;
-  margin-top: -5px;
-  display: block;
-}
-:host /deep/ {
-  .mat-sidenav-content {
-    padding: 0;
-  }
-  md-toolbar + hr, md-card-content + hr {
-    margin-left: -5px;
-    margin-right: -5px;
-  }
-}
-.repo-input-button {
-  padding: 5;
-  margin: 10px;
-  margin-left: 20px;
-}
-
-$mat-toolbar-height-desktop: 64px !default;
-$mat-toolbar-height-mobile-portrait: 56px !default;
-$mat-toolbar-height-mobile-landscape: 48px !default;
-
-.mat-card-top {
-  margin-top: -($mat-toolbar-height-desktop);
-}
-
-
-.mat-card-top {
-  margin-top: -($mat-toolbar-height-desktop);
-}
-
-@media ($mat-xsmall) and (orientation: portrait) {
-  .mat-card-top {
-    margin-top: -($mat-toolbar-height-mobile-portrait);
-  }
-}
-
-@media ($mat-small) and (orientation: landscape) {
-  .mat-card-top {
-    margin-top: -($mat-toolbar-height-mobile-landscape);
-  }
-}
